Rewrite config for use in node-config
It is breaking the knex CLI because the DB config is in the main configuration now and there is no knexfiles. It will be fixed in the following commits.
