--- config/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "rules": {
-    "babel/semi": 2
-  }
-}
