+++ config/console.js
@@ -1,0 +1,4 @@
+module.exports = {
+  'port':          1337,
+  'monitorPrefix': 'development-console'
+}
