--- knexfile.js.docker
@@ -1,83 +1,0 @@
-// Update with your config settings.
-
-try {
-  require('@babel/register');
-} catch (e) {
-  // it's ok. might be already enabled
-}
-
-module.exports = {
-
-  development: {
-    client: 'postgresql',
-    connection: {
-      host:     'postgres',
-      database: 'freefeed',
-      user:     'freefeed',
-      password: 'freefeed'
-    },
-    pool: {
-      min: 2,
-      max: 10
-    },
-    migrations: {
-      tableName: 'knex_migrations'
-    },
-    textSearchConfigName: 'pg_catalog.russian'
-  },
-
-  test: {
-    client: 'postgresql',
-    connection: {
-      host:     'localhost',
-      database: 'freefeed_test',
-      user:     'postgres',
-      password: ''
-    },
-    pool: {
-      min: 2,
-      max: 10
-    },
-    migrations: {
-      tableName: 'knex_migrations'
-    },
-    textSearchConfigName: 'pg_catalog.russian'
-  },
-
-  staging: {
-    client: 'postgresql',
-    connection: {
-      host:     'localhost',
-      database: 'my_db',
-      user:     'username',
-      password: 'password'
-    },
-    pool: {
-      min: 2,
-      max: 10
-    },
-    migrations: {
-      tableName: 'knex_migrations'
-    },
-    textSearchConfigName: 'pg_catalog.russian'
-  },
-
-  production: {
-    client: 'postgresql',
-    connection: {
-      host:     'localhost',
-      database: 'my_db',
-      user:     'username',
-      password: 'password'
-    },
-    pool: {
-      min: 2,
-      max: 10
-    },
-    migrations: {
-      tableName: 'knex_migrations'
-    },
-    textSearchConfigName: 'pg_catalog.russian'
-  }
-
-};
